+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+0.3.1 (2022-12-01)
+------------------
++ Updated fj-core version to 0.8.3
+* Updated fj-doc version to 0.5.5
+* Updated snakeyaml to 1.33
+* Added junit for YamlDocConfig
+* Added [yaml-doc-config sample](src/test/resources/yaml-doc-config.xml)
+* Better logging of info and errors
+
 0.3.0 (2022-11-27)
 ------------------
 + Changed parent pom version (fj-bom) to 0.2.2